feat: fix NotFoundPage.tsx
